% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{abundances}
\alias{abundances}
\alias{ab,}
\alias{otu}
\title{Abundance Matrix from Phyloseq (microbiome)}
\usage{
abundances(x, transform = "identity")
}
\value{
Abundance matrix (OTU x samples).
}
\description{
Retrieves the taxon abundance table from
phyloseq-class object and ensures it is systematically returned as
taxa x samples matrix.
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{internal}
