% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxo_fun.R
\name{add_blast_fun}
\alias{add_blast_fun}
\title{add_blast_fun}
\usage{
add_blast_fun(
  data = data,
  lineage_table = NULL,
  domain = "Bacteria",
  output = "./add_blast/",
  prefix = c("k__", "p__", "c__", "o__", "f__", "g__", "s__")
)
}
\arguments{
\item{data}{Phyloseq object}

\item{lineage_table}{Lineage table from taxonkit software.}

\item{domain}{Domain to use (Bacteria or Fungi)}

\item{output}{Output directory}

\item{prefix}{Prefix for taxonomy to use.}
}
\value{
Phyloseq object with taxonomy table updated.
}
\description{
Add taxonomy information from taxonkit lineage table to phyloseq object.
}
