% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fun.R
\name{aggregate_fun}
\alias{aggregate_fun}
\title{Aggregate}
\usage{
aggregate_fun(
  data = data,
  metacoder = NULL,
  deseq = NULL,
  mgseq = NULL,
  output = "./aggregate_diff/",
  column1 = NULL,
  column2 = NULL,
  verbose = 1,
  rank = "Species",
  comp = "",
  returnval = TRUE
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Aggregate
}
