% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{aggregate_top_taxa}
\alias{aggregate_top_taxa}
\title{aggregate_top_taxa from microbiome package}
\usage{
aggregate_top_taxa(x, top, level)
}
\arguments{
\item{x}{phyloseq object}

\item{top}{Keep the top-n taxa, and merge the rest under the category 'Other'. Instead of top-n numeric this can also be a character vector listing the groups to combine.}

\item{level}{Summarization level (from ‘rank_names(pseq)’)}
}
\description{
aggregate_top_taxa from microbiome package
}
