% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_alpha_fun.R
\name{alphaPlot}
\alias{alphaPlot}
\title{alpha diversité graphique}
\usage{
alphaPlot(data = data, col1 = "", col2 = "", measures = c("Shannon"))
}
\arguments{
\item{data}{A phyloseq object}

\item{col1}{Metadata column name.}

\item{col2}{Metadata column name.}

\item{measures}{"Observed","Shannon","Simpson","InvSimpson"}
}
\value{
A plot.
}
\description{
alpha diversité graphique
}
