% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxo_fun.R
\name{assign_taxo}
\alias{assign_taxo}
\title{Assign Taxo}
\usage{
assign_taxo(
  dada_res = dada_res,
  out = "./idtaxa/",
  id_db = "/home/db/unite/UNITE_idtaxa201706.Rdata",
  verbose = 1,
  confidence = 50,
  returnval = TRUE
)
}
\arguments{
\item{amplicon}{Choose amplipcon "16S" or "ITS"}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Processing DADA2 algorithm on raw sequences, return raw otu table in phyloseq object.
}
