% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxo_fun.R
\name{assign_taxo_fun}
\alias{assign_taxo_fun}
\title{ASV taxonomic assignment with DECIPHER::IDTAXA (assign_taxo)}
\usage{
assign_taxo_fun(
  dada_res = dada_res,
  output = "./idtaxa/",
  id_db = "/PathToDB/UNITE_idtaxa.Rdata",
  confidence = 50,
  verbose = 1,
  returnval = TRUE,
  ncpu = NULL,
  prefix = c("k__", "p__", "c__", "o__", "f__", "g__", "s__"),
  ranks_names = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{output}{Output directory}

\item{id_db}{Vector with list of absolute path to IDTAXA formatted reference database(s).}

\item{confidence}{Bootstrap threshold 0...100}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}

\item{ncpu}{Number of cpus to use.}

\item{prefix}{Vector of prefixes to use.}

\item{ranks_names}{Taxonomy ranks names.}
}
\value{
Return a taxonomy table with multiple ancestor checking and incongruence checking when more than one databases are used.
}
\description{
ASV taxonomic assignment with DECIPHER::IDTAXA (assign_taxo)
}
