% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVenn_fun.R
\name{ASVenn_fun}
\alias{ASVenn_fun}
\title{ASVenn}
\usage{
ASVenn_fun(
  data = data,
  output = "./ASVenn/",
  rank = "ASV",
  column1 = NULL,
  subset = "",
  lvls = "",
  krona = "",
  shared = TRUE
)
}
\arguments{
\item{output}{Output directory}

\item{rank}{Taxonomic rank name, or 'ASV' for ASV level.}

\item{column1}{Factor to test.}

\item{subset}{Subset sample, please provide as c(FACTOR,LEVEL).}

\item{lvls}{Vector list levels of factor to print in venn diagram (max. 5).}

\item{krona}{Krona of exclusive ASV or shared with informed level and others. Must be among levels of column1 argument.}

\item{shared}{shared [TRUE] or exclusive [FALSE] mode.}

\item{dada_res}{Results of dada2_fun()}
}
\value{
Export a venn diagram with corresponding tabulated file.
}
\description{
ASVenn
}
