% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun2}
\alias{bars_fun2}
\title{Barplots plotly}
\usage{
bars_fun2(
  data = data,
  rank = "Genus",
  top = 10,
  Ord1 = NULL,
  Fact1 = NULL,
  relative = TRUE
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{rank}{Rank to output}

\item{top}{Number of top taxa to plot}

\item{Ord1}{Variable used to order sample (X axis)}

\item{Fact1}{Variable used to change X axis tick labels and color categories}

\item{relative}{Plot relative plot (TRUE, default), raw abundance plot (FALSE)}

\item{output}{}
}
\value{
Exports barplots in an interactive plotly community plot
}
\description{
Barplots plotly
}
