% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_soft.R
\name{core_soft_fun}
\alias{core_soft_fun}
\title{core_soft_fun}
\usage{
core_soft_fun(
  data = NULL,
  fact = NULL,
  group = NULL,
  freq = 0.001,
  prev = 0.5,
  rank = "ASV"
)
}
\arguments{
\item{data}{A phyloseq object}

\item{fact}{Factor to test (must be in sample_variables(data))}

\item{group}{Choose which level of the factor, if NULL generate a list for each level.}

\item{freq}{frequence threshold of microbiome::core_members function}

\item{prev}{prevalence threshold of microbiome::core_members function}

\item{rank}{Taxonomy rank.}
}
\description{
Define core microbiome, soft micriome and transitory microbiome.
}
