% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fastq.R
\name{count_fastq_sequences}
\alias{count_fastq_sequences}
\title{count_fastq_sequences}
\usage{
count_fastq_sequences(file_path)
}
\arguments{
\item{file_path}{Path to the fastq file}
}
\value{
The number of sequences in the fastq file
}
\description{
Count the number of sequences in a fastq file.
}
