% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fastq.R
\name{count_seq}
\alias{count_seq}
\title{count_seq}
\usage{
count_seq(path, pattern = ".*R1.*fastq.*")
}
\arguments{
\item{path}{Path to the fastq files directory}

\item{pattern}{Regular expression to match the fastq files}
}
\value{
A vector of sequence counts
}
\description{
Count the number of reads in fastq files, .gz files are allowed.
}
