% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2phyloseq_fun.R
\name{csv2phyloseq_fun}
\alias{csv2phyloseq_fun}
\title{Create Phyloseq object from external datas.}
\usage{
csv2phyloseq_fun(
  otutable = NULL,
  taxtable = NULL,
  seq = NULL,
  metadata = NULL,
  generateTree = FALSE,
  output = "./csv2phyloseq/",
  returnval = TRUE
)
}
\arguments{
\item{otutable}{Path to the tabulated ASV/OTU table or data.frame object.}

\item{taxtable}{Path to the tabulated taxonomy table or data.frame object.}

\item{seq}{to sequence in fasta format or DNAStringSet object.}

\item{metadata}{Path to the tabulated metadata table or data.frame object.}

\item{generateTree}{Boolean to generate the phylogenetic tree.}

\item{output}{Output directory}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a phyloseq object
}
\description{
Create Phyloseq object from external datas.
}
