% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_fun.R
\name{dada2_fun}
\alias{dada2_fun}
\title{DADA2 process}
\usage{
dada2_fun(
  amplicon = "16S",
  path = "",
  outpath = "./dada2_out/",
  f_trunclen = 240,
  r_trunclen = 240,
  f_primer = "GCATCGATGAAGAACGCAGC",
  r_primer = "TCCTCCGCTTWTTGWTWTGC",
  plot = FALSE,
  compress = FALSE,
  verbose = 1
)
}
\arguments{
\item{amplicon}{Choose amplipcon "16S" or "ITS"}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Processing DADA2 algorithm on raw sequences, return raw otu table in phyloseq object.
}
