% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_fun.R
\name{dada_filter}
\alias{dada_filter}
\title{Frequency filter on dada_fun output}
\usage{
dada_filter(dada_res = dada_res, freq = 5e-05)
}
\arguments{
\item{dada_res}{a dada_fun output}

\item{freq}{a frequency threshold to filter rare ASVs}
}
\value{
Filtered dada_fun output
}
\description{
Apply filter to eliminate rare ASVs, allowing to reduce the size of the dataset and to remove potential contaminants.
}
\examples{
\dontrun{
dada_filter(dada_res = dada_res, freq = 0.00005)
}
}
