% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam_fun.R
\name{decontam_fun}
\alias{decontam_fun}
\title{Decontam Function}
\usage{
decontam_fun(
  data = data,
  output = "./decontam_out/",
  number = 4000,
  prev = 2,
  freq = 5e-05,
  column = "type",
  ctrl_identifier = "control",
  spl_identifier = "sample",
  batch = NULL,
  plot = FALSE,
  method = "prevalence",
  threshold = 0.1,
  concentration = NULL,
  verbose = 1,
  domain = TRUE,
  unassigned = FALSE,
  skip = FALSE,
  manual_cont_rank = "Genus",
  manual_cont = NULL,
  returnval = TRUE
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Decontam Function
}
