% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_fun.R
\name{deseq2_fun}
\alias{deseq2_fun}
\title{DESEQ2 Analyse Diff}
\usage{
deseq2_fun(
  data = data,
  output = "./deseq/",
  column1 = "",
  verbose = 1,
  rank = "Species",
  comp = ""
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
DESEQ2 Analyse Diff
}
