% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_alpha_fun.R
\name{diversity_alpha_fun}
\alias{diversity_alpha_fun}
\title{Diversity Alpha}
\usage{
diversity_alpha_fun(
  data = data,
  output = "./plot_div_alpha/",
  column1 = "",
  column2 = "",
  column3 = "",
  supcovs = "",
  measures = c("Observed", "Shannon", "Simpson", "InvSimpson")
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Diversity Alpha
}
