% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_fun.R
\name{diversity_beta_fun}
\alias{diversity_beta_fun}
\title{Diversity Beta}
\usage{
diversity_beta_fun(
  data = data,
  output = "./plot_div_beta/",
  glom = "ASV",
  column1 = "",
  column2 = "",
  covar = "",
  supp = FALSE
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{glom}{Taxonomic rank to agglomerate data (one of rank_names(data) )}

\item{column1}{Column name of main factor to test}

\item{column2}{Column name to split dataset with.}

\item{covar}{One or more factor name to integrate as covariable in permanova. (for multiple covariable, provide as vector)}

\item{supp}{Supplementary plot (jaccard indexes + CCA/RDA ordination)}
}
\value{
Export plots and tests in the output directory.
}
\description{
Provides ordination (PCoA, NMDS) of beta diversity indices (BrayCurtis, UniFrac and WeightedUnifrac) and statistical tests like PERMANOVA and pairwisePERMANOVA.
}
