% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_stamp_fun.R
\name{export_to_stamp_fun}
\alias{export_to_stamp_fun}
\title{Export to stamp}
\usage{
export_to_stamp_fun(data = data, output = "./stamp/", correc = TRUE)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Export 2 text file to use with STAMP
}
