% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_phyloseq_fun.R
\name{generate_phyloseq_fun}
\alias{generate_phyloseq_fun}
\title{Generate Tree}
\usage{
generate_phyloseq_fun(
  otutable = dada_res,
  taxtable = tax.table,
  tree = tree,
  metadata = "",
  output = "./phyloseq/",
  verbose = 1,
  returnval = TRUE
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Generate Tree
}
