% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_fun.R
\name{generate_tree_fun}
\alias{generate_tree_fun}
\title{Generate Tree}
\usage{
generate_tree_fun(dada_res = dada_res, output = "./tree", returnval = TRUE)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Generate Tree
}
