% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_fun.R
\name{get.sample.name}
\alias{get.sample.name}
\title{Get sample name}
\usage{
get.sample.name(fname)
}
\arguments{
\item{fname}{a parameter}
}
\description{
Extract sample names, assuming filenames have format names_miscinfo.fastq
}
\keyword{internal}
