% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fun.R
\name{heatmap_fun}
\alias{heatmap_fun}
\title{Heatmap}
\usage{
heatmap_fun(
  data = data,
  column1 = "",
  output = "./plot_heatmap/",
  rank = "Genus"
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{column1}{Column name of factor to plot with.}

\item{output}{Output directory}

\item{rank}{Taxonomic rank name.}
}
\value{
Export plots and tests in the output directory.
}
\description{
Heatmap generator
}
