% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDeseq2_fun.R
\name{heatmapDeseq2_fun}
\alias{heatmapDeseq2_fun}
\title{heatmap for deseq2_fun output}
\usage{
heatmapDeseq2_fun(desq, phys, var, workingRank)
}
\arguments{
\item{desq}{output of deseq2_fun}

\item{phys}{the phyloseq object used for deseq2 analyse}

\item{var}{Factor to test.}

\item{workingRank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}
}
\value{
Return a list object with heatmap plots.
}
\description{
heatmap for deseq2_fun output
}
