% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDeseq2_fun.R
\name{heatmapDeseq2_funSub}
\alias{heatmapDeseq2_funSub}
\title{heatmap for deseq2_fun output}
\usage{
heatmapDeseq2_funSub(desq, phys, var, workingRank, i)
}
\arguments{
\item{desq}{output of deseq2_fun}

\item{phys}{the phyloseq object used for deseq2 analyse}

\item{var}{Factor to test.}

\item{workingRank}{Taxonomy rank used in deseq2_fun function.}

\item{i}{deseq subset name}
}
\value{
Return a list object with heatmap plots.
}
\description{
heatmap for deseq2_fun output
}
