% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idataxa_assign_fun.R
\name{idTaxa_assign}
\alias{idTaxa_assign}
\title{IDTAXA assign}
\usage{
idTaxa_assign(db_file, dna, asv_names, confidence)
}
\arguments{
\item{amplicon}{Choose amplipcon "16S" or "ITS"}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Processing///
}
