% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fasta_fun.R
\name{idtaxa_assign_fasta_fun}
\alias{idtaxa_assign_fasta_fun}
\title{IDTAXA assign}
\usage{
idtaxa_assign_fasta_fun(
  fasta,
  id_db,
  output = "./assign_fasta/",
  confidence = 50,
  verbose = 1,
  returnval = TRUE,
  ncpu = NULL,
  prefix = c("k__", "p__", "c__", "o__", "f__", "g__", "s__"),
  ranks_names = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{fasta}{Path to fasta file, or DNAStringSet object.}

\item{id_db}{Vector with list of absolute path to IDTAXA formatted reference database(s).}

\item{output}{Output directory}

\item{confidence}{Bootstrap threshold 0...100}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}

\item{ncpu}{Number of cpu to use.}

\item{prefix}{Vector of prefixes to use.}

\item{ranks_names}{Taxonomy ranks names.}
}
\value{
Return a taxonomy table with multiple ancestor checking and incongruence checking when more than one databases are used.
}
\description{
Same function as assign_taxo_fun but use FASTA file as input.
}
