% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxaDB_formatting_functions.R
\name{idtaxa_traindb}
\alias{idtaxa_traindb}
\title{IDTAXA db training (idtaxa training functions)}
\usage{
idtaxa_traindb(
  taxtable = taxtable,
  taxid = taxid,
  seqs = "",
  prunedb = NULL,
  outputDIR = "./",
  outputDBname = "newDB.rdata",
  returnval = FALSE
)
}
\arguments{
\item{taxtable}{Output from check_tax_fun function.}

\item{taxid}{Output from taxid_fun function}

\item{seqs}{FASTA files of reference sequence}

\item{outputDIR}{output directory path}

\item{outputDBname}{db name}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Save rdata files with the formatted idtaxa reference database.
}
\description{
IDTAXA db training (idtaxa training functions)
}
