% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder_fun.R
\name{launch_metacoder}
\alias{launch_metacoder}
\title{Metacoder Differential Analysis function.}
\usage{
launch_metacoder(psobj, min, col, rank, var, title, plot1, signif)
}
\arguments{
\item{psobj}{A phyloseq object.}

\item{min}{Minimum read number.}

\item{col}{A column name.}

\item{rank}{A taxonomy rank. (one of rank_names(data) )}
}
\description{
Metacoder Differential Analysis function.
}
