% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_soft.R
\name{list_venn_fun}
\alias{list_venn_fun}
\title{List generator for venndiagram}
\usage{
list_venn_fun(inlist = NULL, part = "core")
}
\arguments{
\item{inlist}{Output of core_soft_fun() function}

\item{part}{which part of microbiome to output (core, soft or transit)}
}
\description{
Output list to generate venn diagram with VENNFUN function.
}
