% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder_fun.R
\name{metacoder_fun}
\alias{metacoder_fun}
\title{Metacode Analyse Diff}
\usage{
metacoder_fun(
  data = data,
  output = "./metacoder",
  column1 = "",
  column2 = "",
  rank = "Species",
  signif = TRUE,
  plottrees = FALSE,
  min = "1000",
  comp = "all"
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Metacode Analyse Diff
}
