% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeseq_fun.R
\name{metagenomeseq_fun}
\alias{metagenomeseq_fun}
\title{MEtagenomeSeq Analyse Diff}
\usage{
metagenomeseq_fun(
  data = data,
  output = "./metagenomeseq/",
  column1 = "",
  verbose = 1,
  rank = "Species",
  comp = ""
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
MEtagenomeSeq Analyse Diff
}
