% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{mixedsort}
\alias{mixedsort}
\alias{mixedorder}
\title{gtools Order or Sort strings with embedded numbers so that the numbers are in the
correct order}
\usage{
mixedsort(
  x,
  decreasing = FALSE,
  na.last = TRUE,
  blank.last = FALSE,
  numeric.type = c("decimal", "roman"),
  roman.case = c("upper", "lower", "both"),
  scientific = TRUE
)

mixedorder(
  x,
  decreasing = FALSE,
  na.last = TRUE,
  blank.last = FALSE,
  numeric.type = c("decimal", "roman"),
  roman.case = c("upper", "lower", "both"),
  scientific = TRUE
)
}
\arguments{
\item{x}{Vector to be sorted.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
Note that \code{descending=TRUE} reverses the meanings of \code{na.last} and
\code{blanks.last}.}

\item{na.last}{for controlling the treatment of \code{NA} values.  If
\code{TRUE}, missing values in the data are put last; if \code{FALSE}, they
are put first; if \code{NA}, they are removed.}

\item{blank.last}{for controlling the treatment of blank values.  If
\code{TRUE}, blank values in the data are put last; if \code{FALSE}, they
are put first; if \code{NA}, they are removed.}

\item{numeric.type}{either "decimal" (default) or "roman".  Are numeric
values represented as decimal numbers (\code{numeric.type="decimal"}) or as
Roman numerals (\code{numeric.type="roman"})?}

\item{roman.case}{one of "upper", "lower", or "both".  Are roman numerals
represented using only capital letters ('IX') or lower-case letters ('ix')
or both?}

\item{scientific}{logical. Should exponential notation be allowed for numeric values.}
}
\value{
\code{mixedorder} returns a vector giving the sort order of the
input elements. \code{mixedsort} returns the sorted vector.
}
\description{
These functions sort or order character strings containing embedded numbers
so that the numbers are numerically sorted rather than sorted by character
value.  I.e. "Aspirin 50mg" will come before "Aspirin 100mg".  In addition,
case of character strings is ignored so that "a", will come before "B" and
"C".
}
\details{
I often have character vectors (e.g. factor labels), such as compound and
dose, that contain both text and numeric data.  This function is useful for
sorting these character vectors into a logical order.

It does so by splitting each character vector into a sequence of character
and numeric sections, and then sorting along these sections, with numbers
being sorted by numeric value (e.g. "50" comes before "100"), followed by
characters strings sorted by character value (e.g. "A" comes before "B")
\emph{ignoring case} (e.g. 'A' has the same sort order as 'a').

By default, sort order is ascending, empty strings are sorted to the front,
and \code{NA} values to the end.  Setting \code{descending=TRUE} changes the
sort order to descending and reverses the meanings of \code{na.last} and
\code{blank.last}.

Parsing looks for decimal numbers unless \code{numeric.type="roman"}, in
which parsing looks for roman numerals, with character case specified by
\code{roman.case}.
}
\seealso{
\code{\link[base]{sort}}, \code{\link[base]{order}}
}
\author{
Gregory R. Warnes \email{greg@warnes.net}
}
\keyword{internal}
