% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{mtabulate}
\alias{mtabulate}
\title{qdaptools::mtabulate Tabulate Frequency Counts for Multiple Vectors}
\usage{
mtabulate(vects)
}
\arguments{
\item{vects}{A \code{\link[base]{vector}}, \code{\link[base]{list}}, or 
\code{\link[base]{data.frame}} of named/unnamed vectors.}
}
\value{
Returns a \code{\link[base]{data.frame}} with columns equal to 
number of unique elements and the number of rows equal to the the original 
length of the \code{\link[base]{vector}}, \code{\link[base]{list}}, or 
\code{\link[base]{data.frame}} (length equals ncols in 
\code{\link[base]{data.frame}}).  If list of vectors is named 
these will be the rownames of the dataframe.
}
\description{
Similar to \code{\link[base]{tabulate}} that works on multiple vectors.
}
\references{
\url{https://stackoverflow.com/a/9961324/1000343}
}
\seealso{
\code{\link[base]{tabulate}}, \code{\link[qdapTools]{counts2list}}
}
\author{
Joran Elias and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{frequency}
\keyword{internal}
\keyword{tabulate}
