% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2cyto_fun.R
\name{phy2cyto_fun}
\alias{phy2cyto_fun}
\title{Export files for cytoscape.}
\usage{
phy2cyto_fun(
  data = data,
  output = "./cytoscape/",
  column1 = NULL,
  repl = NULL,
  verbose = 1
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{column1}{Factor to test.}

\item{repl}{Replicates column.}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}
}
\value{
Export files ready to use with Cytoscape
}
\description{
Export files for cytoscape.
}
\references{
https://cytoscape.org/
}
