% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2tsv_fun.R
\name{phy2tsv_fun}
\alias{phy2tsv_fun}
\title{Phyloseq to TSV}
\usage{
phy2tsv_fun(data = data, output = "./tsv_table/", rank = "ASV")
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
EXPORT otu table with taxonomy and sequences, + metadata in separated file.
}
