% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_fun.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot beta diversity}
\usage{
plot_beta(psobj, rank, col, path, var = "total", cov = covar)
}
\arguments{
\item{psobj}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{col}{A column name from metadata (among sample_variables(data))}

\item{path}{Output directory}

\item{var}{Name of analysis for output customization}

\item{cov}{One or more column names from metadata to treat as a covariable (among sample_variables(data)), provided as comma separated vector.}
}
\description{
Plot beta diversity
}
