% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxaDB_formatting_functions.R
\name{prune_db_fun}
\alias{prune_db_fun}
\title{Prune DB}
\usage{
prune_db_fun(taxtable = taxtable, seqs = "", prunedb = 10, outputDIR = "./")
}
\arguments{
\item{taxtable}{data.frame}

\item{seqs}{path to fasta file or readDNAStringSet}

\item{prunedb}{maximum number of sequences per unique taxa.}
}
\value{
List with taxonomy table and corresponding sequences.
}
\description{
Prune DB
}
