% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{rarefaction}
\alias{rarefaction}
\title{Rarefaction plotly}
\usage{
rarefaction(data = data, col = NULL, step = 100)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{col}{The metadata column name you want to split the graph.}

\item{step}{The number of steps you want your rarefaction plots to be done.}
}
\value{
A plotly graph.
}
\description{
Rarefaction plotly
}
