% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_non_ascii}
\alias{remove_non_ascii}
\title{Remove non-ASCII characters from a string}
\usage{
remove_non_ascii(text)
}
\arguments{
\item{text}{A string}
}
\description{
Remove non-ASCII characters from a string
}
\keyword{internal}
