% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_fastx.R
\name{subset_fastx}
\alias{subset_fastx}
\title{subset_fastx}
\usage{
subset_fastx(
  path = NULL,
  format = "fastq",
  outformat = "fastq",
  output = "./subset_fastq/",
  nbseq = 10000,
  ncores = 3,
  compress = FALSE,
  verbose = FALSE,
  random = FALSE,
  seed = NULL
)
}
\arguments{
\item{path}{Path to the fastq files directory}

\item{format}{input format, fasta or fastq format are allowed.}

\item{outformat}{output format, fasta or fastq format are allowed.}

\item{output}{Path to the output directory}

\item{nbseq}{Number of sequences to output per fastq files; if NULL no subset.}

\item{ncores}{Number of CPU used to process.}

\item{compress}{set on TRUE to get .fastx.gz compressed files}

\item{verbose}{Print information while processing}

\item{random}{Pick randomly 'nbseq' sequences.}

\item{seed}{Fix seed with numeric to reproduce random picking (eg. 123)}
}
\value{
Output fastx files in the output directory
}
\description{
Allows subset fastq or fasta files at a given threshold. This function can convert fastq to fasta.
}
