% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2tsv_fun.R
\name{taxa_counts}
\alias{taxa_counts}
\title{taxa_counts}
\usage{
taxa_counts(data = NULL, except = NULL)
}
\arguments{
\item{data}{a phyloseq object}

\item{except}{exclude taxa (eg. unclassified|unknown)}
}
\value{
List with counts and details
}
\description{
Function to count the number of taxa at each rank on a phyloseq object
}
