% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxaDB_formatting_functions.R
\name{taxid_fun}
\alias{taxid_fun}
\title{Taxid processing (idtaxa training functions)}
\usage{
taxid_fun(taxtable = taxtable, output = "./taxid.txt")
}
\arguments{
\item{taxtable}{Output from check_tax_fun function.}

\item{output}{output path}
}
\value{
Return the taxid object.
}
\description{
Taxid processing (idtaxa training functions)
}
