% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{top_taxa}
\alias{top_taxa}
\title{Top Taxa (microbiome)}
\usage{
top_taxa(x, n = ntaxa(x))
}
\arguments{
\item{x}{phyloseq object}

\item{n}{Number of top taxa to return (default: all)}
}
\value{
Character vector listing the top taxa
}
\description{
Return n most abundant taxa (based on total abundance over
all samples), sorted by abundance
}
\keyword{internal}
