% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata_fun.R
\name{update_metadata_fun}
\alias{update_metadata_fun}
\title{Update metadata}
\usage{
update_metadata_fun(
  data = data,
  output = "./updated_physeq/",
  metadata = NULL,
  returnval = TRUE
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{metadata}{Tabulated metadata file path.}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a phyloseq object with updated metadata.
}
\description{
Update metadata
}
