% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVenn_fun.R
\name{VENNFUN}
\alias{VENNFUN}
\title{VENNFUN}
\usage{
VENNFUN(
  TF = TF,
  mode = 1,
  TITRE = TITRE,
  output = "./",
  refseq1 = NULL,
  alltax = NULL,
  ggplotmode = FALSE
)
}
\arguments{
\item{TF}{list containing vectors to plot.}

\item{mode}{= 1: length(TF)<=5, mode = 2 5<length(TF)<7}

\item{TITRE}{Plot title.}

\item{output}{Output path.}

\item{refseq1}{Reference sequences.}

\item{alltax}{Taxonomy table.}

\item{ggplotmode}{if TRUE plot the Venn diagram using ggplot}
}
\value{
Exports a venn diagram with corresponding tabulated file.
}
\description{
Plotting Venn Function and table of shared feature generation
}
