% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fun.R
\name{aggregate_fun}
\alias{aggregate_fun}
\title{Aggregate}
\usage{
aggregate_fun(
  data = data,
  metacoder = NULL,
  deseq = NULL,
  mgseq = NULL,
  output = "./aggregate_diff/",
  column1 = NULL,
  column2 = NULL,
  verbose = 1,
  rank = "Species",
  comp = "",
  returnval = TRUE
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{metacoder}{Path to the metacoder CSV file}

\item{deseq}{Path to deseq results folder}

\item{mgseq}{Path to metagenomeseq results folder}

\item{output}{Output directory}

\item{column1}{Column name of factor to test}

\item{column2}{Column name on which table were splitted.}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{rank}{Taxonomic rank name, or 'ASV' for ASV level.}

\item{comp}{Comparison to test. Comma separated and comparisons are informed with a tilde (A~C,A~B,B~C). If empty, test all combination.}
}
\value{
Export final CSV files, barplot with top significant ASV and Venn Digramm.
}
\description{
Aggregate results of the tree differential analysis methods in one single file.
}
