% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_alpha_fun.R
\name{alphaPlot}
\alias{alphaPlot}
\title{alpha diversité graphique}
\usage{
alphaPlot(data = data, col1 = "", col2 = "", measures = c("Shannon"))
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{col1}{Metadata column name (from sample_variables(data)).}

\item{col2}{Metadata column name (from sample_variables(data)).}

\item{measures}{Indices among "Observed","Shannon","Simpson","InvSimpson"}
}
\value{
A plot.
}
\description{
alpha diversité graphique
}
