% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxo_fun.R
\name{assign_taxo_fun}
\alias{assign_taxo_fun}
\title{ASV taxonomic assignment with DECIPHER::IDTAXA (assign_taxo)}
\usage{
assign_taxo_fun(
  dada_res = dada_res,
  output = "./idtaxa/",
  id_db = "/PathToDB/UNITE_idtaxa.Rdata",
  confidence = 50,
  verbose = 1,
  returnval = TRUE,
  ncpu = NULL
)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{output}{Output directory}

\item{id_db}{Vector with list of absolute path to IDTAXA formatted reference database(s) (up to 2 databases).}

\item{confidence}{Bootstrap threshold 0...100}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}

\item{ncpu}{Number of cpus to use.}
}
\value{
Return a taxonomy table with multiple ancestor checking and incongruence checking when 2 databases are used.
}
\description{
ASV taxonomic assignment with DECIPHER::IDTAXA (assign_taxo)
}
