% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVenn_fun.R
\name{ASVenn_fun}
\alias{ASVenn_fun}
\title{ASVenn}
\usage{
ASVenn_fun(
  data = data,
  output = "./ASVenn/",
  rank = "ASV",
  column1 = NULL,
  subset = "",
  lvls = "",
  krona = "",
  shared = TRUE
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{output}{Output directory}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{column1}{Column name of factor to test (among sample_variables(data))}

\item{subset}{Subset sample, please provide as c(FACTOR,LEVEL).}

\item{lvls}{Vector comma separated list levels of factor to print in venn diagram (max. 5).}

\item{krona}{Krona of exclusive ASV or shared with informed level and others. Must be among levels of column1 argument.}

\item{shared}{shared [TRUE] or exclusive [FALSE] mode.}
}
\value{
Export a venn diagram with corresponding tabulated file.
}
\description{
Function to create Venn Diagram of shared features
}
