% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun}
\alias{bars_fun}
\title{Barplots community}
\usage{
bars_fun(
  data = data,
  bar = TRUE,
  compo1 = TRUE,
  output = "./plot_bar/",
  column1 = "",
  column2 = "",
  sname = FALSE,
  num = 10,
  rare = NULL,
  rank = "Genus"
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Barplots community
}
