% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun}
\alias{bars_fun}
\title{Barplots plotly}
\usage{
bars_fun(
  data = data,
  rank = "Genus",
  top = 10,
  Ord1 = NULL,
  Fact1 = NULL,
  split = FALSE,
  relative = TRUE,
  autoorder = TRUE,
  ylab = "Abundance",
  outfile = "plot_compo.html",
  verbose = TRUE
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{top}{Number of top taxa to plot}

\item{Ord1}{Variable used to order sample (X axis) or split the barplot if split = TRUE}

\item{Fact1}{Variable used to change X axis tick labels and color (when split = FALSE)}

\item{split}{if TRUE make a facet_wrap like grouped by Ord1 (default FALSE)}

\item{relative}{Plot relative (TRUE, default) or raw abundance plot (FALSE)}

\item{autoorder}{Order xaxis with gtools::mixedorder function (TRUE).}

\item{ylab}{Y axis title ("Abundance")}

\item{outfile}{Output html file.}
}
\value{
Returns barplots in an interactive plotly community plot
}
\description{
Barplots plotly
}
