% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun}
\alias{bars_fun}
\title{Barplots community}
\usage{
bars_fun(
  data = data,
  bar = TRUE,
  compo1 = TRUE,
  output = "./plot_bar/",
  column1 = "",
  column2 = "",
  sname = FALSE,
  num = 10,
  rare = NULL,
  rank = "Genus"
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{bar}{Plotting bar plot with raw reads number.}

\item{output}{Output directory}

\item{column1}{Column name of factor used to sort sample}

\item{column2}{Column name of factor used to split barplot}

\item{sname}{Change sample.id by the corresponding factor levels in graph.}

\item{num}{Number of top taxon to display.}

\item{rare}{Column name for splitting rare curves.}

\item{rank}{Taxonomic rank name. You can provide multiple ranks seperated by comma.}

\item{compo}{Plotting with relative composition.}
}
\value{
Export barplots in an html file.
}
\description{
Barplots community
}
