% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun}
\alias{bars_fun}
\title{Barplots plotly}
\usage{
bars_fun(
  data = data,
  rank = "Genus",
  top = 10,
  Ord1 = NULL,
  sample_labels = FALSE,
  split = FALSE,
  split_sid_order = FALSE,
  relative = TRUE,
  levelsOrder = TRUE,
  autoorder = TRUE,
  ylab = "Abundance",
  outfile = "plot_compo.html",
  verbose = TRUE,
  pal = NULL
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{top}{Number of top taxa to plot}

\item{Ord1}{Variable used to order sample (X axis) or split the barplot if split = TRUE}

\item{sample_labels}{If true, x axis labels are sample IDS, if false labels displayed are levels from Ord1 argument. Ignored if split = TRUE (FALSE)}

\item{split}{if TRUE make a facet_wrap like grouped by Ord1 (default FALSE)}

\item{split_sid_order}{keep sample_ids order from metadata for splitted plots.}

\item{relative}{Plot relative (TRUE, default) or raw abundance plot (FALSE)}

\item{levelsOrder}{Ordering samples according to levels of Ord1 variable. If FALSE, order from the metadata table is kept.}

\item{autoorder}{Automatic ordering xaxis labels based on Ord1 factor levels with gtools::mixedorder function (TRUE).}

\item{ylab}{Y axis title ("Abundance")}

\item{outfile}{Output html file.}

\item{verbose}{Boolean verbosity level.}

\item{pal}{Color palette.}
}
\value{
Returns barplots in an interactive plotly community plot
}
\description{
Barplots plotly
}
