% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxaDB_formatting_functions.R
\name{check_tax_fun}
\alias{check_tax_fun}
\title{Check taxonomy table (idtaxa training functions)}
\usage{
check_tax_fun(
  taxtable = taxtable,
  output = NULL,
  rank = 7,
  verbose = 3,
  returnval = TRUE
)
}
\arguments{
\item{taxtable}{Output from fill_tax_table function.}

\item{output}{output path}

\item{rank}{Deepest taxonomy rank at which correction begin (7 for Species, 6 for Genus etc...).}

\item{verbose}{verbose mode}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return the same taxonomy without incongruence like multiple ancestors, by replacing problematic taxonomy by the most common.
}
\description{
Check taxonomy table (idtaxa training functions)
}
