% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2phyloseq_fun.R
\name{csv2phyloseq_fun}
\alias{csv2phyloseq_fun}
\title{Create Phyloseq object from external datas.}
\usage{
csv2phyloseq_fun(
  otutable = NULL,
  taxtable = NULL,
  seq = NULL,
  metadata = NULL,
  generateTree = FALSE,
  output = "./csv2phyloseq/",
  returnval = TRUE
)
}
\arguments{
\item{otutable}{Tabulated ASVtable file path.}

\item{taxtable}{Tabulated taxonomy table file path.}

\item{seq}{Tabulated sequence file path.}

\item{metadata}{Tabulated metadata file path.}

\item{generateTree}{Boolean to generate the phylogenetic tree.}

\item{output}{Output directory}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a phyloseq object
}
\description{
Create Phyloseq object from external datas.
}
