% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_fun.R
\name{dada2_fun}
\alias{dada2_fun}
\title{DADA2 process (dada2_fun())}
\usage{
dada2_fun(
  path = "",
  outpath = "./dada2_out/",
  cutadapt = FALSE,
  maxEE = 5,
  f_trunclen = 240,
  r_trunclen = 240,
  dadapool = "pseudo",
  f_primer = "GCATCGATGAAGAACGCAGC",
  r_primer = "TCCTCCGCTTWTTGWTWTGC",
  plot = FALSE,
  compress = FALSE,
  extension = "_R1.fastq",
  verbose = 1,
  torrent_single = FALSE,
  returnval = TRUE,
  paired = TRUE,
  trim_l = 15,
  trim_r = 0,
  orient_torrent = NULL,
  n_cpu = 6
)
}
\arguments{
\item{path}{Read files folder path}

\item{outpath}{output .Rdata file name}

\item{cutadapt}{Use of cutadapt to trim primers based on their sequences, f_ and r_primer are needed (ambiguous nucleotides allowed)}

\item{maxEE}{Maximum expected error in reads (in filterAndTrim function) (default: 5)}

\item{f_trunclen}{Forward read truncate length (only for paired end 16S)}

\item{r_trunclen}{Reverse read truncate length (only for paired end 16S)}

\item{dadapool}{option for dada function (FALSE, TRUE or "pseudo"), default is "pseudo". See ? dada.}

\item{f_primer}{Forward primer sequence (mandatory if cutadapt = TRUE),  5' -> 3'.}

\item{r_primer}{Reverse primer sequence (mandatory if cutadapt = TRUE), 5' -> 3'.}

\item{plot}{Plot all test or not}

\item{compress}{Reads files are compressed (.gz)}

\item{extension}{Paired reads extension files for R1. (default: _R1.fastq), use "_R1.fastq.gz" for compressed files.}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{torrent_single}{Boolean to choose between Illumina Paired End SOP or Torrent Single End SOP. default: FALSE}

\item{returnval}{Boolean to return values in console or not.}

\item{paired}{Boolean for Illumina Paired End Reads.}

\item{trim_l}{Trim left size, to trim forward primer without cutadapt.}

\item{trim_r}{Trim right size, to trim reverse primer without cutadapt.}

\item{orient_torrent}{Forward primer sequence to orient all reads to same strand (only unambiguous nucleotides).}

\item{n_cpu}{Number of CPU to use in multithread.}
}
\value{
Return raw otu table in phyloseq object and export it in an Rdata file.
}
\description{
Processing DADA2 algorithm on raw sequences (IonTorrent or Illumina / single or paired end), return raw otu table with representative sequence of ASV.
}
\examples{
See https://forgemia.inra.fr/umrf/ranomaly/-/wikis/home#dada2-usage-according-to-raw-data-type

}
