% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam_fun.R
\name{decontam_fun}
\alias{decontam_fun}
\title{Decontam Function using decontam package}
\usage{
decontam_fun(
  data = data,
  domain = TRUE,
  output = "./decontam_out/",
  number = 4000,
  prev = 2,
  freq = 5e-05,
  column = "",
  ctrl_identifier = "control",
  spl_identifier = "sample",
  batch = NULL,
  plot = FALSE,
  method = "prevalence",
  threshold = 0.1,
  concentration = NULL,
  verbose = 1,
  unassigned = FALSE,
  skip = FALSE,
  manual_cont_rank = "Genus",
  manual_cont = NULL,
  krona = FALSE,
  returnval = TRUE
)
}
\arguments{
\item{data}{a phyloseq object output from generate_phyloseq_fun()}

\item{domain}{16S region or ITS region (Bacteria or Fungi).}

\item{output}{Output directory}

\item{number}{Minimum number of reads per sample.}

\item{prev}{Minimum prevalence of an ASV in samples to be kept.}

\item{freq}{Minimum ASV frequence on overall samples.}

\item{column}{Column name from sample_variables(data) for type of sample (control or sample). If informed, function filters control samples.}

\item{ctrl_identifier}{Idendifier name for controls.}

\item{spl_identifier}{Idendifier name for samples.}

\item{batch}{Batch column name for independent contaminant identification.}

\item{plot}{Plot all test.}

\item{method}{Method for contaminant identification. (frequency, prevalence, combined, both, either).}

\item{threshold}{Threshold for DECONTAM prevalence filtering.}

\item{concentration}{Column name for ADN concentration.}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{unassigned}{Unassigned kingdom or phylum fitering.}

\item{skip}{Skip decontam step and process basic filters (depth per samples, prevalence, frequence).}

\item{manual_cont_rank}{Rank of taxa to remove, inform 'ASV' to remove ASV.}

\item{manual_cont}{Comma separated list of Genus to remove (eg. "g__Enterococcus,g__Cellulosimicrobium,g__Serratia").}

\item{krona}{Export krona plot}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a decontaminated phyloseq object.
}
\description{
Decontam Function using decontam package
}
