% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_alpha_fun.R
\name{diversity_alpha_fun}
\alias{diversity_alpha_fun}
\title{Diversity Alpha}
\usage{
diversity_alpha_fun(
  data = data,
  output = "./plot_div_alpha/",
  column1 = "",
  column2 = "",
  column3 = "",
  supcovs = "",
  measures = c("Observed", "Shannon", "Simpson", "InvSimpson"),
  verbose = TRUE
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{output}{Output directory}

\item{column1}{Column name of first factor to test (covariable in ANOVA).}

\item{column2}{Column name of second factor to test (last factor in ANOVA).}

\item{column3}{Column name of subjects in case of repeated mesures (mixed model)}

\item{supcovs}{One or more supplementary covariables to test in anova (provided as comma separated vector)}

\item{measures}{Diversity indices (provided as comma separated vector)}

\item{verbose}{Boolean}
}
\value{
Return plots and tests as list and export them in the output directory.
}
\description{
Provides boxplots with multiples diversity indices and statistical tests like ANOVA with post hoc test and non parametric Wilcoxon tests.
}
