% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_fun.R
\name{diversity_beta_fun}
\alias{diversity_beta_fun}
\title{Diversity Beta}
\usage{
diversity_beta_fun(
  data = data,
  output = "./plot_div_beta/",
  glom = "ASV",
  column1 = "",
  column2 = "",
  covar = ""
)
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Diversity Beta
}
