% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_light.R
\name{diversity_beta_light}
\alias{diversity_beta_light}
\title{Diversity Beta}
\usage{
diversity_beta_light(
  psobj,
  rank = "ASV",
  col = NULL,
  cov = NULL,
  dist0 = "bray",
  ord0 = "MDS",
  output = "./plot_div_beta/",
  tests = TRUE
)
}
\arguments{
\item{psobj}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{col}{A metadata column (among sample_variables(data)).}

\item{cov}{Covariable names comma separated vector.}

\item{dist0}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{ord0}{Currently supported method options are: c("DCA", "CCA", "RDA", "CAP", "DPCoA", "NMDS", "MDS", "PCoA")}

\item{output}{The output file directory.}

\item{tests}{Whether to compute tests or not (TRUE/FALSE)}
}
\value{
Return specific plots and tests in list and output them in the output directory.
}
\description{
Provides ordination (PCoA, NMDS) of beta diversity indices (BrayCurtis, UniFrac and WeightedUnifrac) and statistical tests like PERMANOVA and pairwisePERMANOVA.
}
