% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_light.R
\name{diversity_beta_light}
\alias{diversity_beta_light}
\title{Diversity Beta}
\usage{
diversity_beta_light(
  psobj,
  rank = "ASV",
  col = NULL,
  cov = NULL,
  dist0 = "bray",
  ord0 = "MDS",
  output = "./plot_div_beta/",
  tests = TRUE
)
}
\arguments{
\item{output}{Output directory}

\item{data}{output from decontam or generate_phyloseq}

\item{glom}{Taxonomic rank to agglomerate data (one of rank_names(data) )}

\item{column1}{Column name of main factor to test}

\item{covar}{One or more factor name to integrate as covariable in permanova. (for multiple covariable, provide as vector)}

\item{column2}{Column name to split dataset with.}

\item{supp}{Supplementary plot (jaccard indexes + CCA/RDA ordination)}
}
\value{
Export plots and tests in the output directory.
}
\description{
Provides ordination (PCoA, NMDS) of beta diversity indices (BrayCurtis, UniFrac and WeightedUnifrac) and statistical tests like PERMANOVA and pairwisePERMANOVA.
}
