% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_stamp_fun.R
\name{export_to_stamp_fun}
\alias{export_to_stamp_fun}
\title{Export to stamp}
\usage{
export_to_stamp_fun(data = data, output = "./stamp/", correc = FALSE)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{correc}{If TRUE, correct metadata to replace most common special characters (eg. é -> e), save the new file in meta_stampOK.tsv.}
}
\value{
Export 2 text files ready to use with STAMP.
}
\description{
Export 2 text file to use with STAMP
}
\references{
https://beikolab.cs.dal.ca/software/STAMP
}
