% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxaDB_formatting_functions.R
\name{fill_tax_fun}
\alias{fill_tax_fun}
\title{Fill taxonomy table (idtaxa training functions)}
\usage{
fill_tax_fun(taxtable = taxtable)
}
\arguments{
\item{taxtable}{Taxonomy table in tabulated format.}
}
\value{
Return the same taxonomy without empty field, last known ranks are informed.
}
\description{
Fill taxonomy table (idtaxa training functions)
}
