% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_phyloseq_fun.R
\name{generate_phyloseq_fun}
\alias{generate_phyloseq_fun}
\title{Generate Tree}
\usage{
generate_phyloseq_fun(
  dada_res = dada_res,
  taxtable = tax.table,
  tree = tree,
  metadata = "",
  output = "./phyloseq/",
  verbose = 1,
  returnval = TRUE
)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{taxtable}{Results of assign_taxo_fun()}

\item{tree}{Results of generate_tree_fun()}

\item{metadata}{Path of metadata file (tab separated with header)}

\item{output}{Output directory}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
Generate Tree
}
