% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_phyloseq_fun.R
\name{generate_phyloseq_fun}
\alias{generate_phyloseq_fun}
\title{Generate Phyloseq Object}
\usage{
generate_phyloseq_fun(
  dada_res = dada_res,
  tax.table = tax.table,
  tree = NULL,
  metadata = "",
  output = "./phyloseq/",
  verbose = 1,
  returnval = TRUE
)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{tax.table}{Results of assign_taxo_fun()}

\item{tree}{Results of generate_tree_fun()}

\item{metadata}{Path of metadata file (tab separated with header)}

\item{output}{Output directory}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a complete phyloseq object.
}
\description{
Generate Phyloseq Object
}
