% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_fun.R
\name{generate_tree_fun}
\alias{generate_tree_fun}
\title{Generate Phylogenetic Tree}
\usage{
generate_tree_fun(dada_res = dada_res, output = "./tree", returnval = TRUE)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{output}{Output directory}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a formatted tree object ready to use in phyloseq.
}
\description{
Generate Phylogenetic Tree
}
