% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_fun.R
\name{generate_tree_fun}
\alias{generate_tree_fun}
\title{Generate Phylogenetic Tree}
\usage{
generate_tree_fun(
  dada_res = NULL,
  psobj = NULL,
  output = "./tree",
  returnval = TRUE,
  n_cpus = 6,
  verbose = 1
)
}
\arguments{
\item{dada_res}{Results of dada2_fun()}

\item{psobj}{Phyloseq object with sequences.}

\item{output}{Output directory.}

\item{returnval}{Boolean to return values in console or not.}

\item{n_cpus}{Number of cpus to use.}

\item{verbose}{Verbosity level.}
}
\value{
Return a formatted tree object ready to use in phyloseq.
}
\description{
Generate Phylogenetic Tree
}
