% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fun.R
\name{heatmap_fun}
\alias{heatmap_fun}
\title{Heatmap}
\usage{
heatmap_fun(
  data = data,
  column1 = "",
  top = 20,
  output = "./plot_heatmap/",
  rank = "Species"
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{column1}{Column name of factor to plot with.}

\item{top}{Number of top features to plot.}

\item{output}{Output directory}

\item{rank}{Taxonomic rank name.}
}
\value{
Generate an heatmap with top taxa (html plotly version in output directory)
}
\description{
Generate an heatmap with top taxa.
}
