% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fun.R
\name{heatmap_fun}
\alias{heatmap_fun}
\title{Heatmap}
\usage{
heatmap_fun(
  data = data,
  column1 = "",
  top = 20,
  output = "./plot_heatmap/",
  rank = "Species"
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{column1}{Column name of factor to plot with (among sample_variables(data)).}

\item{top}{Number of top features to plot.}

\item{output}{Output directory}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}
}
\value{
Generate an heatmap with top taxa (html plotly version in output directory)
}
\description{
Generate an heatmap with top taxa.
}
