% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fun.R
\name{heatmap_fun}
\alias{heatmap_fun}
\title{Heatmap}
\usage{
heatmap_fun(
  data = data,
  column1 = "",
  top = NULL,
  output = "./plot_heatmap/",
  rank = "Species",
  norm = "TSS",
  legend = "Abundance",
  clust = TRUE
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{column1}{Column name of factor to plot with (among sample_variables(data)).}

\item{top}{If not NULL, only this number of top features are plotted. Non "top" taxa abundances are aggregated in a new taxa named "Other".}

\item{output}{Output directory}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{norm}{Normalization method ("TSS" or "VST"), needs a phyloseq object with raw abundance.}

\item{legend}{Legend title ("Abundance")}

\item{clust}{If TRUE, taxa are reordered with standard clustering.}
}
\value{
Generate an heatmap with top taxa (html plotly version in output directory)
}
\description{
Generate an heatmap with top taxa.
}
