% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDeseq2_fun.R
\name{heatmapDeseq2_fun}
\alias{heatmapDeseq2_fun}
\title{heatmap for deseq2_fun output}
\usage{
heatmapDeseq2_fun(desq, phys, var, workingRank, output = "./deseq/heatmap/")
}
\arguments{
\item{desq}{output of deseq2_fun}

\item{phys}{the phyloseq object used for deseq2 analyse}

\item{var}{Factor to test.}

\item{workingRank}{Taxonomy rank used in deseq2_fun function.}

\item{output}{Output directory}
}
\value{
Return a list object with heatmap plots and data matrix.
}
\description{
heatmap for deseq2_fun output
}
