% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtaxa_assign_fun.R
\name{idTaxa_assign}
\alias{idTaxa_assign}
\title{IDTAXA assign}
\usage{
idTaxa_assign(db_file, dna, asv_names, confidence)
}
\arguments{
\item{db_file}{Path to IDTAXA formatted database file.}

\item{dna}{A ‘DNAStringSet’ of unaligned sequences.}

\item{asv_names}{sequences IDs in same order.}

\item{confidence}{Bootstrap threshold 0...100}
}
\value{
return taxonomic assignment of given sequences.
}
\description{
Additionnal function for assign_taxo and assign_fasta function.
}
