% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fasta_fun.R
\name{idtaxa_assign_fasta_fun}
\alias{idtaxa_assign_fasta_fun}
\title{IDTAXA assign}
\usage{
idtaxa_assign_fasta_fun(
  fasta,
  id_db,
  output = "./assign_fasta/",
  confidence = 50,
  verbose = 1,
  returnval = TRUE
)
}
\arguments{
\item{fasta}{Path to fasta file, or DNAStringSet object.}

\item{id_db}{Vector with list of absolute path to IDTAXA formatted reference database(s) (up to 2 databases).}

\item{output}{Output directory}

\item{confidence}{Bootstrap threshold 0...100}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{returnval}{Boolean to return values in console or not.}
}
\value{
Return a taxonomy table with multiple ancestor checking and incongruence checking when 2 databases are used.
}
\description{
Same function as assign_taxo_fun but use FASTA file as input.
}
