% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder_fun.R
\name{metacoder_fun}
\alias{metacoder_fun}
\title{Metacoder Differential Analysis function.}
\usage{
metacoder_fun(
  data = data,
  output = "./metacoder",
  column1 = "",
  column2 = "",
  rank = "Species",
  signif = TRUE,
  plottrees = FALSE,
  min = "1000",
  comp = "all"
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{column1}{Column name of factor to test}

\item{column2}{Column name to split dataset with.}

\item{rank}{Taxonomic rank to agglomerate data (one of rank_names(data) )}

\item{signif}{Plot only significant.}

\item{plottrees}{Plot trees (long treatments).}

\item{min}{Minimum number of reads for a taxa to be represented.}

\item{comp}{Comma separated list of comparison to test. Comparisons are informed with a tilde (A~C,A~B,B~C). If empty, test all combination}
}
\value{
Export plots and CSV file listing significant differentialy abundant ASVs.
}
\description{
Metacoder Differential Analysis function.
}
