% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeseq_fun.R
\name{metagenomeseq_fun}
\alias{metagenomeseq_fun}
\title{MEtaGenomeSeq Differential Analysis function.}
\usage{
metagenomeseq_fun(
  data = data,
  output = "./metagenomeseq/",
  column1 = "",
  verbose = 1,
  rank = "Species",
  comp = ""
)
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{column1}{Column name of factor to test}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{rank}{Taxonomic rank to agglomerate data (one of rank_names(data) )}

\item{comp}{Comma separated list of comparison to test. Comparisons are informed with a tilde (A~C,A~B,B~C). If empty, test all combination}
}
\value{
Export CSV files with significant differentialy abondant ASV.
}
\description{
MEtaGenomeSeq Differential Analysis function.
}
