% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeseq_fun.R
\name{metagenomeseq_fun}
\alias{metagenomeseq_fun}
\title{MEtaGenomeSeq Differential Analysis function.}
\usage{
metagenomeseq_fun(
  data = data,
  output = "./metagenomeseq/",
  column1 = "",
  verbose = 1,
  rank = "Species",
  comp = ""
)
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{output}{Output directory}

\item{column1}{Column name of factor to test (among sample_variables(data))}

\item{verbose}{Verbose level. (1: quiet, 3: verbal)}

\item{rank}{Taxonomy rank to merge features that have same taxonomy at a certain taxonomic rank (among rank_names(data), or 'ASV' for no glom)}

\item{comp}{Comma separated list of comparison to test. Comparisons are informed with a tilde (A~C,A~B,B~C). If empty, test all combination}
}
\value{
Returns list with table of features and plots for each comparison. Export CSV files with significant differentialy abondant ASV.
}
\description{
MEtaGenomeSeq Differential Analysis function.
}
