% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2tsv_fun.R
\name{phy2tsv_fun}
\alias{phy2tsv_fun}
\title{Phyloseq to TSV}
\usage{
phy2tsv_fun(data = data, output = "./tsv_table/", rank = "ASV")
}
\arguments{
\item{data}{output from decontam or generate_phyloseq}

\item{output}{Output directory}

\item{rank}{Taxonomic rank used to agglomerate otu table.}
}
\value{
Export tabulated otu table and metadata
}
\description{
Export otu table with taxonomy and sequences, + metadata in separated file.
}
