% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_fun.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot beta diversity}
\usage{
plot_beta(psobj, rank, col, path, var = "total", cov = covar)
}
\arguments{
\item{psobj}{A phyloseq object}

\item{rank}{A taxonomic rank}

\item{col}{A column name from metadata}

\item{path}{The output directory path}

\item{cov}{A column name from metadata to treat as a covariable}
}
\description{
Plot beta diversity
}
