% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_fun.r
\name{plsda_fun}
\alias{plsda_fun}
\title{PLSDA from MixOmics package}
\usage{
plsda_fun(data = data, output = "./plsda/", column1 = "", rank = "Species")
}
\arguments{
\item{data}{A phyloseq object}

\item{output}{Output directory}

\item{column1}{Factor to test.}

\item{rank}{Taxonomy rank (among rank_names(data))}
}
\value{
Return a list object with plots, and loadings table.
}
\description{
Multivariate methods for discriminant analysis.
}
