% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_table_fun.R
\name{split_table_fun}
\alias{split_table_fun}
\title{SPLIT phyloseq object}
\usage{
split_table_fun(data = data, output = "./", column1 = "type")
}
\arguments{
\item{dada_res}{output from dada2_fun}
}
\value{
Return raw otu table in phyloseq object.
}
\description{
SPLIT phyloseq object
}
