% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_table_fun.R
\name{split_table_fun}
\alias{split_table_fun}
\title{Split phyloseq object}
\usage{
split_table_fun(data = data, output = "./", column1 = "")
}
\arguments{
\item{data}{a phyloseq object (output from decontam or generate_phyloseq)}

\item{output}{Output directory}

\item{column1}{Column name of factor to split phyloseq object with. (among sample_variables(data))}
}
\value{
Export new folders with splited phyloseq Robjects
}
\description{
Split phyloseq object
}
